requirements.txt
appdirs==1.4.4
argcomplete==1.12.0
argon2-cffi==20.1.0
attrs==20.1.0
audioread==2.1.8
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.2
chainer==7.7.0
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.2.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Distance==0.1.3
entrypoints==0.3
espnet-tts-frontend==0.0.3
filelock==3.0.12
Flask==1.1.1
future==0.18.2
g2p-en==2.1.0
gdown==3.12.2
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
idna==2.10
inflect==4.1.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jaconv==0.2.4
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
kaldiio==2.15.1
kiwisolver==1.2.0
librosa==0.8.0
llvmlite==0.31.0
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
nltk==3.5
notebook==6.1.3
numba==0.48.0
numpy==1.19.1
packaging==20.4
pandocfilters==1.4.2
parallel-wavegan==0.4.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pooch==1.1.1
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.13.0
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pypinyin==0.39.0
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
PyYAML==5.3.1
pyzmq==19.0.2
regex==2020.7.14
requests==2.24.0
resampy==0.2.2
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
tensorboardX==2.1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
torch==1.6.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
typing-extensions==3.7.4.2
Unidecode==1.1.1
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
xmltodict==0.12.0
yq==2.10.1
